+++ index.html
@@ -3,40 +3,23 @@
 <head>
   <meta charset="UTF-8" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Panel de Administración - KiosKeys</title>
-  <link rel="stylesheet" href="/assets/css/style.css">
+  <title>Panel de Administración - Kioskeys</title>
+  <link rel="stylesheet" href="/style.css">
   <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.0.0/css/all.min.css">
   <script src="https://cdn.jsdelivr.net/npm/sweetalert2@11"></script>
   <script src="https://ucarecdn.com/libs/widget/3.x/uploadcare.full.min.js"></script>
-  <style>
-    :root {
-      --primary-blue: #007BFF; /* Define color si no está */
-    }
-
-    body {
-      font-family: Arial, sans-serif;
-      background-color: #f8f9fa;
-    }
-    
-    nav .logo img {
-      max-height: 50px;
-    }
-    
-    /* Otros estilos heredados */
-  </style>
+  <link rel="stylesheet" href="/src/admin/css/admin.css">
 </head>
 <body>
-  <header>
-    <nav aria-label="Main navigation">
-      <div class="logo">
-        <a href="/"><img src="/assets/images/logo.svg" alt="KiosKeys Logo"></a>
-      </div>
-      <div class="nav-links">
-        <a href="/">Inicio</a>
-        <a href="/admin" class="active">Panel Admin</a>
-      </div>
-    </nav>
-  </header>
+  <nav>
+    <div class="logo">
+      <a href="/"><img src="/logo.svg" alt="Kioskeys" data-logo></a>
+    </div>
+    <div class="nav-links">
+      <a href="/">Inicio</a>
+      <a href="/admin" class="active">Panel Admin</a>
+    </div>
+  </nav>
 
   <main>
     <section class="hero">
@@ -47,7 +30,7 @@
     </section>
 
     <div class="admin-container">
-      <nav class="admin-nav" aria-label="Subnavigation">
+      <nav class="admin-nav">
         <a href="#brands" class="active" data-section="brands">Marcas</a>
         <a href="#products" data-section="products">Productos</a>
         <a href="#categories" data-section="categories">Categorías</a>
@@ -58,75 +41,101 @@
         <div class="section-title">
           <h2>Gestión de Marcas</h2>
         </div>
+        
         <form id="brandForm">
           <input type="hidden" name="id">
+          
           <div class="form-group">
-            <label for="brandLogo">Logo de la Marca</label>
+            <label>Logo de la Marca</label>
             <input type="hidden" role="uploadcare-uploader" data-images-only="true">
             <input type="hidden" name="logo" id="brandLogo">
-            <img id="logoPreview" src="/assets/images/placeholder-logo.png" class="preview-image" alt="Logo Preview">
+            <img id="logoPreview" src="/placeholder-logo.png" class="preview-image">
           </div>
+
           <div class="form-group">
-            <label for="brandName">Nombre de la Marca</label>
-            <input type="text" name="name" id="brandName" required>
+            <label>Nombre de la Marca</label>
+            <input type="text" name="name" required>
           </div>
+
           <div class="form-group">
-            <label for="brandType">Tipo</label>
-            <select name="type" id="brandType" required>
+            <label>Tipo</label>
+            <select name="type" required>
               <option value="car">Automóvil</option>
               <option value="motorcycle">Motocicleta</option>
               <option value="commercial">Comercial</option>
             </select>
           </div>
+
           <button type="submit" class="btn btn-primary">
             <i class="fas fa-save"></i> Guardar Marca
           </button>
         </form>
+
         <div id="brandsList" class="grid">
-          <!-- Las marcas se renderizan aquí -->
+          <!-- Brands will be rendered here -->
         </div>
       </section>
 
-      <section id="products" class="admin-section" hidden>
+      <section id="products" class="admin-section" style="display: none;">
         <div class="section-title">
           <h2>Gestión de Productos</h2>
         </div>
+        <!-- Products content -->
       </section>
 
-      <!-- Otras secciones -->
+      <section id="categories" class="admin-section" style="display: none;">
+        <div class="section-title">
+          <h2>Gestión de Categorías</h2>
+        </div>
+        <!-- Categories content -->
+      </section>
+
+      <section id="settings" class="admin-section" style="display: none;">
+        <div class="section-title">
+          <h2>Configuración</h2>
+        </div>
+        <!-- Settings content -->
+      </section>
     </div>
   </main>
 
   <script type="module">
-    import { BrandsManager } from '/src/admin/js/brands.js';
+    import { BrandsManager } from '/src/admin/js/managers/BrandsManager.js';
     import { db } from '/src/admin/js/db.js';
 
+    // Initialize database and managers
     document.addEventListener('DOMContentLoaded', async () => {
       try {
         await db.open();
         window.brandsManager = new BrandsManager();
 
+        // Handle navigation
         const navLinks = document.querySelectorAll('.admin-nav a');
         const sections = document.querySelectorAll('.admin-section');
 
         navLinks.forEach(link => {
           link.addEventListener('click', (e) => {
             e.preventDefault();
+            
+            // Update active state
             navLinks.forEach(l => l.classList.remove('active'));
             link.classList.add('active');
+
+            // Show selected section
             const sectionId = link.getAttribute('href').substring(1);
             sections.forEach(section => {
-              section.hidden = section.id !== sectionId;
+              section.style.display = section.id === sectionId ? 'block' : 'none';
             });
           });
         });
+
       } catch (error) {
+        console.error('Error initializing admin panel:', error);
         Swal.fire({
           icon: 'error',
           title: 'Error',
           text: 'Error al inicializar el panel de administración'
         });
-        console.error('Error:', error);
       }
     });
   </script>
